added typescript dev dependacy
