+++ package.json
@@ -6,12 +6,21 @@
   "scripts": {
     "android": "npx react-native run-android",
     "start": "npx react-native start",
-    "adb-reverse": "adb reverse tcp:8081 tcp:8081"
+    "adb-reverse": "adb reverse tcp:8081 tcp:8081",
+    "tsc": "npx tsc",
+    "metro-reset": "watchman watch-del-all; rm -rf /tmp/metro-*; npm run start --reset-cache"
   },
   "author": "vaidyakhil",
   "license": "ISC",
   "dependencies": {
     "react": "^17.0.2",
-    "react-native": "^0.66.4" 
+    "react-native": "^0.66.4"
+  },
+  "devDependencies": {
+    "@types/jest": "^27.4.0",
+    "@types/react": "^17.0.38",
+    "@types/react-native": "^0.66.12",
+    "@types/react-test-renderer": "^17.0.1",
+    "typescript": "^4.5.4"
   }
 }
